Moved Iterator & Enum to base.py
